requirements.txt
asgiref==3.7.2
colorclass==2.2.2
crispy-bootstrap5==2023.10
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.2
django-appconf==1.0.6
django-ckeditor-5==0.2.11
django-crispy-forms==2.1
django-filter==23.5
django-imagekit==5.0.0
django-import-export==3.3.7
django-meta==2.4.2
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pilkit==3.0
pillow==10.2.0
pip-upgrader==1.4.15
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
terminaltables==3.1.10
typing_extensions==4.9.0
tzdata==2024.1
Unidecode==1.3.8
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
